Fix documentation linking or missing images and add new utilities to avoid repeating ourselves
