+++ spark/src/main/kotlin/com/adevinta/spark/components/buttons/ButtonSuccess.kt
@@ -25,7 +25,6 @@ package com.adevinta.spark.components.buttons
 import androidx.compose.foundation.interaction.Interaction
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -39,7 +38,6 @@ import com.adevinta.spark.components.text.Text
 import com.adevinta.spark.icons.IdentityCardOutline
 import com.adevinta.spark.icons.LeboncoinIcons
 import com.adevinta.spark.icons.SparkIcon
-import com.adevinta.spark.tokens.disabled
 
 /**
  * Used for positive confirmations like "Approve" or "Mark as sold" where the outcome is beneficial.
@@ -72,16 +70,9 @@ internal fun SparkButtonSuccess(
     atEnd: Boolean = false,
     content: @Composable RowScope.() -> Unit,
 ) {
-    val containerColor = SparkTheme.colors.surface
-    val contentColor = SparkTheme.colors.success
-    val disabledContainerColor = containerColor.disabled
-    val disabledContentColor = contentColor.disabled
-
-    val colors = ButtonDefaults.outlinedButtonColors(
-        containerColor = containerColor,
-        disabledContainerColor = disabledContainerColor,
-        contentColor = contentColor,
-        disabledContentColor = disabledContentColor,
+    val colors = SparkButtonDefaults.outlinedButtonColors(
+        containerColor = SparkTheme.colors.surface,
+        contentColor = SparkTheme.colors.success,
     )
     BaseSparkButton(
         onClick = onClick,
@@ -103,6 +94,8 @@ internal fun SparkButtonSuccess(
 
 /**
  * Used for positive confirmations like "Approve" or "Mark as sold" where the outcome is beneficial.
+ *
+ * ![Button Success](https://leboncoin.github.io/spark-android/images/com.adevinta.spark.buttons_NewButtonDocumentationScreenshots_buttonSuccess.png)
  *
  * @param onClick Will be called when the user clicks the button
  * @param modifier Modifier to be applied to the button
